+++ src/app/SideNav.svelte
@@ -58,11 +58,6 @@
       <i class="fa fa-search mr-2" /> Search
     </a>
   </li>
-  <li class="cursor-pointer">
-    <a class="block px-4 py-2 transition-all hover:bg-accent hover:text-white" href="/scan">
-      <i class="fa fa-qrcode mr-2" /> Scan
-    </a>
-  </li>
   <li
     class={cx("relative", {
       "cursor-pointer": $canPublish,