+++ src/app/shared/PersonBadge.svelte
@@ -1,4 +1,5 @@
 <script>
+  import cx from "classnames"
   import {Link} from "svelte-routing"
   import {killEvent} from "src/util/html"
   import {displayPerson} from "src/util/nostr"
@@ -10,14 +11,14 @@
 </script>
 
 {#if inert}
-  <span class="relative z-10 flex items-center gap-2">
+  <span class={cx($$props.class, "relative z-10 flex items-center gap-2")}>
     <PersonCircle {person} />
     <span class="text-lg font-bold">{displayPerson(person)}</span>
   </span>
 {:else}
   <Link
     to={routes.person(person.pubkey)}
-    class="relative z-10 flex items-center gap-2"
+    class={cx($$props.class, "relative z-10 flex items-center gap-2")}
     on:click={killEvent}>
     <PersonCircle {person} />
     <span class="text-lg font-bold">{displayPerson(person)}</span>