+++ src/app/views/Notifications.svelte
@@ -4,7 +4,12 @@
   import {onMount} from "svelte"
   import {fly} from "svelte/transition"
   import {navigate} from "svelte-routing"
-  import {now, timedelta, formatTimestampAsDate, createScroller} from "src/util/misc"
+  import {
+    now,
+    formatTimestampAsDate,
+    formatTimestampAsLocalISODate,
+    createScroller,
+  } from "src/util/misc"
   import {findReplyId} from "src/util/nostr"
   import Spinner from "src/partials/Spinner.svelte"
   import Tabs from "src/partials/Tabs.svelte"
@@ -30,7 +35,7 @@
     // Sort by rounded timestamp so we can group reactions to the same parent
     return reverse(
       sortBy(
-        e => Math.round(e.created_at / timedelta(1, "hour")).toString() + findReplyId(e),
+        e => formatTimestampAsLocalISODate(e.created_at) + findReplyId(e),
         user.applyMutes(t.all())
       )
     )