+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [x] Register url handler for web+nostr and use that for sharing
 - [x] Combine search and scan pages
+- [x] Clean up notifications page
 
 # 0.2.28