+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [x] Register url handler for web+nostr and use that for sharing
 - [x] Combine search and scan pages
 - [x] Clean up notifications page
+- [x] Add note share page
 
 # 0.2.28